AI-Backend/pyproject.toml
[project]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = [{name = "prathit2003",email = "arcius777@gmail.com"}]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "cloudinary (>=1.42.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "torch (==2.5.0)",
    "torchvision (==0.20.0)",
    "fastai (>=2.7.18,<3.0.0)",
    "llvmlite (==0.42.0)",
    "realesrgan (>=0.3.0,<0.4.0)",
    "requests (>=2.32.3,<3.0.0)",
    "aiohttp (>=3.11.12,<4.0.0)",
    "onnxruntime (>=1.20.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "boto3 (>=1.38.15,<2.0.0)",
]

[tool.poetry]
packages = [{include = "poetry_demo", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
